server/requirements.txt
# Pipecat and Daily dependencies
pipecat-ai[silero,openai,google]==0.0.90
daily-python==0.19.9
audioop-lts==0.2.2

# FastAPI and server dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9
requests==2.31.0

# AI dependencies
google-generativeai>=0.8.5
openai>=1.50.2
python-dotenv==1.0.0

# Database dependencies
sqlalchemy==2.0.44
alembic==1.16.5
mako==1.3.10
